+++ report/correct/classes/report.php
@@ -303,7 +303,7 @@ class report extends default_report {
                 $element['downloadlink'] = $link->out();
                 $element['documentname'] = $queue->filename;
                 $element['queueid'] = $queue->id;
-                $element['numberofpages'] = sizeof($queuefilesmatrix) - 1;
+                $element['numberofpages'] = sizeof($queuefilesmatrix);
                 $date = new DateTime();
                 $date->setTimestamp(intval($queue->timecreated));
                 $element['importdate'] = userdate($date->getTimestamp());
